Change authentication logic and add Profile activity
Create additional authentication process on app start
that allows to show different screens depending on user flag. 
If the user is new and doesn't exist in database then 
the signup screen is displayed, otherwise the login screen is
displayed. The logic is implemented using SharedPreferences.

Add Login activity to authenticate an existing user.